kiteconnect::connect

Struct KiteConnect

pub struct KiteConnect {
    api_key: String,
    access_token: String,
    session_expiry_hook: Option<fn()>,
    client: Client,
}

Main KiteConnect struct for interacting with Kite Connect API

Fields
api_key: Stringaccess_token: Stringsession_expiry_hook: Option<fn()>client: Client
Implementations

impl KiteConnect

pub(crate) fn build_url( &self, path: &str, param: Option<Vec<(&str, &str)>>, ) -> Url

Constructs url for the given path and query params

pub fn new(api_key: &str, access_token: &str) -> Self

Constructor

async fn raise_or_return_json(&self, resp: Response) -> Result<JsonValue>

Helper method to raise or return json response for async responses

pub fn set_session_expiry_hook(&mut self, method: fn())

Sets an expiry hook method for this instance

pub fn session_expiry_hook(&self) -> Option<fn()>

Gets the session expiry hook

pub fn set_access_token(&mut self, access_token: &str)

Sets an access token for this instance

pub fn access_token(&self) -> &str

Gets the access token for this instance

pub fn login_url(&self) -> String

Returns the login url

async fn compute_checksum(&self, input: &str) -> Result<String>

Compute checksum for authentication - different implementations for native vs WASM

pub async fn generate_session( &mut self, request_token: &str, api_secret: &str, ) -> Result<JsonValue>

Request for access token

pub async fn invalidate_access_token( &self, access_token: &str, ) -> Result<Response>

Invalidates the access token

pub async fn renew_access_token( &mut self, access_token: &str, api_secret: &str, ) -> Result<JsonValue>

Request for new access token

pub async fn invalidate_refresh_token( &self, refresh_token: &str, ) -> Result<Response>

Invalidates the refresh token

pub async fn margins(&self, segment: Option<String>) -> Result<JsonValue>

Return the account balance and cash margin details for a particular segment

pub async fn profile(&self) -> Result<JsonValue>

Get user profile details

pub async fn holdings(&self) -> Result<JsonValue>

Get all holdings

pub async fn positions(&self) -> Result<JsonValue>

Get all positions

pub async fn place_order( &self, variety: &str, exchange: &str, tradingsymbol: &str, transaction_type: &str, quantity: &str, product: Option<&str>, order_type: Option<&str>, price: Option<&str>, validity: Option<&str>, disclosed_quantity: Option<&str>, trigger_price: Option<&str>, squareoff: Option<&str>, stoploss: Option<&str>, trailing_stoploss: Option<&str>, tag: Option<&str>, ) -> Result<JsonValue>

Place an order

pub async fn modify_order( &self, order_id: &str, variety: &str, quantity: Option<&str>, price: Option<&str>, order_type: Option<&str>, validity: Option<&str>, disclosed_quantity: Option<&str>, trigger_price: Option<&str>, parent_order_id: Option<&str>, ) -> Result<JsonValue>

Modify an open order

pub async fn cancel_order( &self, order_id: &str, variety: &str, parent_order_id: Option<&str>, ) -> Result<JsonValue>

Cancel an order

pub async fn exit_order( &self, order_id: &str, variety: &str, parent_order_id: Option<&str>, ) -> Result<JsonValue>

Exit a BO/CO order

pub async fn orders(&self) -> Result<JsonValue>

Get a list of orders

pub async fn order_history(&self, order_id: &str) -> Result<JsonValue>

Get the list of order history

pub async fn trades(&self) -> Result<JsonValue>

Get all trades

pub async fn order_trades(&self, order_id: &str) -> Result<JsonValue>

Get all trades for a specific order

pub async fn convert_position( &self, exchange: &str, tradingsymbol: &str, transaction_type: &str, position_type: &str, quantity: &str, old_product: &str, new_product: &str, ) -> Result<JsonValue>

Modify an open position product type

pub async fn mf_orders(&self, order_id: Option<&str>) -> Result<JsonValue>

Get all mutual fund orders or individual order info

pub async fn trigger_range( &self, transaction_type: &str, instruments: Vec<&str>, ) -> Result<JsonValue>

Get the trigger range for a list of instruments

pub async fn instruments(&self, exchange: Option<&str>) -> Result<JsonValue>

Get instruments list

pub async fn mf_instruments(&self) -> Result<JsonValue>

Get mutual fund instruments list